Remove bogus escaping of brackets in email bodies
The extra `\` in front of `[` and `]` are not needed. Look at the other templates that work just fine. It was setting a bad example when everyone copy pastes these other templates.
    
This was done with a search and replace of the regex `\\+\]` with `]` and the similar for `[` on all .md and .markdown files. I went through to double check that it worked fine, I didn't see any mistakes.