Render email body as plaintext not HTML
Before, we were passing the email body through the `markdownify` filter, which was a mistake. This made it so that if you copy and pasted the HTML from the webpage you would get something different from the "copy to clipboard" button, which just copied the body as is. The email body is simply plaintext, it's *not* markdown!

This converts all newlines to `<br>` tags, resulting in something that both looks good (nearly identical as with the `<p>` tags from `markdownify`) and can be copied and pasted exactly as originally written.

We now need to escape the content, in case someone puts something that's valid HTML (e.g. `<YOUR NAME HERE>`). I think this is necessary, I don't think we can trust our email reviewers to catch all cases that might break this. I also think that this is sufficient, but perhaps I'm missing something that could break things?

I'm not sure what this means for Siteleaf users? I do know that escaped content such as `\[` now appears exactly like that, and I think Siteleaf creating the escape. However, now that I think about it, whenever the copyToClipboard button is pressed, before you would get `\[` because through that route the content is not markdownified! So `\[` is bad, and a separate problem.

Next step is to ensure people aren't treating the body as markdown. I'll DM the several people who have been doing this (found via git blame) so that they know to stop.