Fix newlines in all email templates
To get over how

```
[YOUR NAME]
[ADDRESS]
[PHONE NUMBER]
[EMAIL ADDRESS]
```

was getting turned into

```
[YOUR NAME] [ADDRESS] [PHONE NUMBER] [EMAIL ADDRESS]
```

once markdownified and displayed in HTML, people were hacking it by adding extra newlines in places. Now, since we are properly not treating the body as markdown, this should get reformatted to plaintext. It will now appear *exactly* as it does here.

A spelling mistake was fixed in there too.

Next step is to ensure this can't happen again. I'll DM the several people who have been doing this (found via git blame) so that they know to stop.